# refactor: move type_coercion to analyzer