… benchmarks/expected-plans/q17.txt
… datafusion/optimizer/src/analyzer/type_coercion.rs