@jackwener 2023-04-05
This PR contains a fix for type coercion of subquery. I will polish it in following PR.
This fix is important, because it will move `cast` from `expression` into `subplan`. It means that we don't `cast expression` in `eval expression` and we do cast before `eval expression`.
But look like it just a little help for performance.

```sql
--- before
tpch q17
cargo run --release --bin tpch -- benchmark datafusion --iterations 5 --path ./data --format tbl --query 17 --batch-size 4096
Query 17 iteration 0 took 5233.1 ms and returned 1 rows
Query 17 iteration 1 took 4940.8 ms and returned 1 rows
Query 17 iteration 2 took 5160.2 ms and returned 1 rows
Query 17 iteration 3 took 5315.6 ms and returned 1 rows
Query 17 iteration 4 took 4967.7 ms and returned 1 rows
Query 17 avg time: 5123.48 ms
```

```sql
--- after
tpch q17
Query 17 iteration 0 took 4789.5 ms and returned 1 rows
Query 17 iteration 1 took 4785.2 ms and returned 1 rows
Query 17 iteration 2 took 4791.5 ms and returned 1 rows
Query 17 iteration 3 took 5051.4 ms and returned 1 rows
Query 17 iteration 4 took 4817.7 ms and returned 1 rows
Query 17 avg time: 4847.07 ms
```