@jackwener 2023-04-05
slt run fail, but it's strange. cc @alamb @melgenek 

```sql
[dates.slt] Running query: "select i_item_desc
from test
where d3_date > d2_date + INTERVAL '1 days';"
Error: statement is expected to fail with error:
[dates.slt] Running query: "select i_item_desc
	DataFusion error: Error during planning: Timestamp(Nanosecond, Some("+00:00")) + Utf8 can't be evaluated because there isn't a common type to coerce the types to
from test
but got error:
where d3_date > d2_date + INTERVAL '5 days';"
	DataFusion error: Error during planning: Timestamp(Nanosecond, Some("+00:00")) + Utf8 can't be evaluated because there isn't a common type to coerce the types to
```