@mingmwang 2023-04-06
I will take a look tomorrow. 

@jackwener One thing need to confirm with you is the `Error` behavior. 
Originally the optimizer rules might return an error, and the logical plan optimizer will skip the failed rule by default and will
not fail the SQL.  But I think for Analyzer rules, if the rule return an `Error`, I think the SQL should failure immediately. 
For the `type_coercion` rule, the rule itself is to ensure the correctness, I think if there is error, the SQL should failure.

#5831

```rust
       /// When set to true, the logical plan optimizer will produce warning
        /// messages if any optimization rules produce errors and then proceed to the next
        /// rule. When set to false, any rules that produce errors will cause the query to fail
        pub skip_failed_rules: bool, default = true
```